+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] Q2 2022
+ - factory reset:
+   - report new status "in_progress"
+ - user consent:
+   - report general consent status
+
 ## [0.3.0] Q2 2022
  - added user consent handling for "iot-hub-device-update"
  - bumped to latest azure-iot-sdk 0.5.5